// client/src/types/movie.ts
// These are the results that come from every movie; not sure if
// I will use all or exclude a few that likely won't be used anyways
export type Movie = {
  adult: false;
  backdrop_path: string;
  genre_ids: Array<number>;
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: false;
  vote_average: number;
  vote_count: number;
};